Moved snackbar component to its own folder
