+++ LokiDashboard/layouts/default.vue
@@ -77,7 +77,7 @@
 
 <script>
 import { mapActions, mapGetters } from "vuex";
-import TheSnackBar from "@/components/TheSnackBar";
+import TheSnackBar from "@/components/notifications/TheSnackBar";
 export default {
   components: {
     TheSnackBar
